% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centered_kernel_matrix_euclidean.R
\name{centered_kernel_matrix_euclidean}
\alias{centered_kernel_matrix_euclidean}
\title{Compute a Centered fBM Kernel Matrix for Euclidean data}
\usage{
centered_kernel_matrix_euclidean(
  eval_points_1,
  eval_points_2,
  centering_grid,
  hurst_coef,
  dimension
)
}
\arguments{
\item{eval_points_1}{Matrix or numeric vector. First set of evaluation points
(e.g., x).}

\item{eval_points_2}{Matrix or numeric vector. Second set of evaluation points
(e.g., x').}

\item{centering_grid}{Matrix or numeric vector. Grid used for kernel centering
(e.g., z1, ..., zn).}

\item{hurst_coef}{Numeric scalar. The Hurst coefficient.}

\item{dimension}{Integer. Dimensionality of the input space.}
}
\value{
A numeric matrix of kernel evaluations.
}
\description{
Computes the centered fractional Brownian motion (fBM) kernel matrix for
Euclidean, order, or graph data.
}
