% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_measure_euclidean.R
\name{get_base_measures_euclidean}
\alias{get_base_measures_euclidean}
\title{Compute Base Measure Weights for Euclidean data}
\usage{
get_base_measures_euclidean(
  samples,
  dimension = 1,
  boundaries,
  parallel_workers = max(1L, parallel::detectCores(logical = TRUE) - 1L)
)
}
\arguments{
\item{samples}{A numeric vector or matrix of sample points. For 1D, a vector; for 2D, a matrix/data.frame with two columns.}

\item{dimension}{Integer. Dimensionality of the sample: either 1 or 2.}

\item{boundaries}{A numeric vector (length 2) for 1D defining min/max bounds, or a 2x2 matrix for 2D, each row defining min/max bounds of each axis.}

\item{parallel_workers}{Integer. Number of involved cores in the calculation.}

\item{method}{String specifying the method of base measure calculation for ordered data.}
}
\value{
A numeric vector of base measure weights.
}
\description{
Computes the base measure weights for either 1D (using grid midpoints) or 2D (using Voronoi tessellation).
}
\examples{
get_base_measures_euclidean(sample = runif(100), boundaries = c(0,1), dimension = 1)
get_base_measures_euclidean(sample = matrix(runif(200), ncol=2), boundaries = matrix(c(0,1,0,1), nrow=2), dimension = 2)
}
