% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_euclidean.R
\name{get_dens_wo_grid_euclidean}
\alias{get_dens_wo_grid_euclidean}
\title{Compute Normalized Densities at Sampled Points Without a Grid for euclidean data.}
\usage{
get_dens_wo_grid_euclidean(
  centered_kernel_mat_samples,
  samples,
  base_measure_weights,
  dimension,
  lambda,
  weight_vec
)
}
\arguments{
\item{centered_kernel_mat_samples}{Centered kernel matrix at sampled points.}

\item{samples}{Sample locations.}

\item{base_measure_weights}{Vector of base measure weights (used for normalizing constant if dim > 1).}

\item{dimension}{Dimensionality of the data (1 or higher).}

\item{lambda}{Scalar lambda parameter.}

\item{weight_vec}{Vector of weights.}
}
\value{
Normalized density values at sampled points.
}
\description{
Compute Normalized Densities at Sampled Points Without a Grid for euclidean data.
}
