% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_euclidean.R
\name{get_weights_euclidean}
\alias{get_weights_euclidean}
\title{Estimate Weights Using the Barzilai-Borwein Method (Optimized with Rcpp) for
Euclidean data.}
\usage{
get_weights_euclidean(
  lambda,
  tau,
  centered_kernel_mat_samples,
  samples,
  base_measure_weights,
  dimension,
  prior_var_prob,
  print_trace = FALSE
)
}
\arguments{
\item{tau}{A scalar representing the tau parameter.}

\item{centered_kernel_mat_samples}{A matrix representing the centered kernel at sampled points.}

\item{samples}{A numeric vector of sampled points.}

\item{base_measure_weights}{A numeric vector of base measures for sample points.}

\item{dimension}{A scalar representing the dimension of sample.}

\item{prior_var_prob}{Logical; if TRUE, the variance of prior is proportional to probability itself and a hyper-parameter, else it is only proportional to a hyper-parameter.}

\item{print_trace}{Logical; if TRUE, prints progress updates.}

\item{lambdas}{A scalar representing the lambda parameter.}
}
\value{
A numeric vector of estimated weights.
}
\description{
This function estimates the weight vector using the Barzilai-Borwein method,
a numerical approach to solve nonlinear systems of equations. It optimizes
performance by leveraging Rcpp for computation and adaptive sub-sampling for
better initial values in the Barzilai-Borwein method.
}
