% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_s_function.R
\name{get_s_function}
\alias{get_s_function}
\title{Compute the \code{s} Function from Weights and Kernel Matrix}
\usage{
get_s_function(
  weight_vec,
  lambda,
  tau,
  centered_kernel_mat_samples,
  samples,
  base_measure_weights,
  dimension,
  data_type,
  prior_var_prob = TRUE
)
}
\arguments{
\item{weight_vec}{A numeric vector of weights.}

\item{lambda}{A scalar regularization parameter.}

\item{tau}{A scalar penalty parameter.}

\item{centered_kernel_mat_samples}{Centered kernel matrix at sampled points (n × n).}

\item{samples}{A numeric matrix of sampled values.}

\item{base_measure_weights}{A numeric vector of base measure weights (same length as \code{samples}).}

\item{dimension}{A scalar indicating the dimension of the data.}

\item{data_type}{String specifying the data type. One of "euclidean", "order", or "graph". The default is "euclidean".}

\item{prior_var_prob}{Logical; if TRUE, the variance of prior is proportional to probability itself and a hyper-parameter, else it is only proportional to a hyper-parameter.}
}
\value{
A numeric vector representing the output of the s-function.
}
\description{
Wrapper for the Rcpp function \code{get_s_function} that computes a vector used in density-based optimization.
}
